more reorganisation of UI and constants files
