+++ qcp/gui/components/combo_box_component.py
@@ -15,10 +15,11 @@ from typing import List
 
 from PySide6 import QtCore, QtWidgets
 from qcp.gui.components import AbstractComponent
-from qcp.gui.constants import (ALGORITHM_LAYOUTS, COMBO_BOX_LAYOUT_MAPPING,
-                               COMBO_BOX_NAME, GROVER_LAYOUT,
-                               LABEL_TEXT_FORMAT, LABEL_TITLE,
-                               LAYOUT_COMBO_MAPPING)
+from qcp.gui.components.constants import (ALGORITHM_LAYOUTS,
+                                          COMBO_BOX_LAYOUT_MAPPING,
+                                          COMBO_BOX_NAME, GROVER_LAYOUT,
+                                          LABEL_TEXT_FORMAT, LABEL_TITLE,
+                                          LAYOUT_COMBO_MAPPING)
 
 
 class ComboBoxComponent(AbstractComponent):
